refactor: StrategyIssueConfigManager 迁移至 bkmonitor/models/issue.py，后台 alarm_backends 不再持有配置写逻辑
